Add handleClick method to Board

+++ Game/Board.h
@@ -18,6 +18,6 @@ private:
 	void draw(sf::RenderTarget & target, sf::RenderStates states) const;
 public:
 	explicit Board(sf::Font & font);
-	void handleClick(float x, float y);
+	void handleClick(int x, int y);
 	sf::Vector2f getSize() const;
 };
